Commentaires ajoutés + Page Index arrangée

+++ Projet/Search-Find-Job/supprimer-annonce.php
@@ -7,10 +7,13 @@ if(!isset($_SESSION))
 session_start();
 }
   $annonce =  GetAnnonceInfo($_GET['idA']);
+  if(!empty($annonce[6])&&!empty($annonce[7])&&!empty($annonce[8]))
   $image = $annonce[6].$annonce[7].".".$annonce[8];
     if(DeleteAnnonce($_GET['idA'],$_GET['idU']))
     {       
+      if(!empty($annonce[6])&&!empty($annonce[7])&&!empty($annonce[8]))
           unlink($image);
+          
           header('location: annonces.php');
     }
     else
